+++ markets/spot-market/contracts/modules/AtomicOrderModule.sol
@@ -23,10 +23,10 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function buyExactOut(
         uint128 marketId,
-        uint synthAmount,
-        uint maxUsdAmount,
+        uint256 synthAmount,
+        uint256 maxUsdAmount,
         address referrer
-    ) external override returns (uint usdAmountCharged, OrderFees.Data memory fees) {
+    ) external override returns (uint256 usdAmountCharged, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
         spotMarketFactory.validateMarket(marketId);
 
@@ -45,7 +45,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
 
         spotMarketFactory.usdToken.transferFrom(msg.sender, address(this), usdAmountCharged);
 
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
@@ -67,10 +67,10 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function buy(
         uint128 marketId,
-        uint usdAmount,
-        uint minAmountReceived,
+        uint256 usdAmount,
+        uint256 minAmountReceived,
         address referrer
-    ) external override returns (uint synthAmount, OrderFees.Data memory fees) {
+    ) external override returns (uint256 synthAmount, OrderFees.Data memory fees) {
         return buyExactIn(marketId, usdAmount, minAmountReceived, referrer);
     }
 
@@ -79,10 +79,10 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function buyExactIn(
         uint128 marketId,
-        uint usdAmount,
-        uint minAmountReceived,
+        uint256 usdAmount,
+        uint256 minAmountReceived,
         address referrer
-    ) public override returns (uint synthAmount, OrderFees.Data memory fees) {
+    ) public override returns (uint256 synthAmount, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
         spotMarketFactory.validateMarket(marketId);
 
@@ -102,7 +102,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
             revert InsufficientAmountReceived(minAmountReceived, synthAmount);
         }
 
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
@@ -124,7 +124,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function quoteBuyExactIn(
         uint128 marketId,
-        uint usdAmount
+        uint256 usdAmount
     ) external view override returns (uint256 synthAmount, OrderFees.Data memory fees) {
         SpotMarketFactory.load().validateMarket(marketId);
 
@@ -142,7 +142,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function quoteBuyExactOut(
         uint128 marketId,
-        uint synthAmount
+        uint256 synthAmount
     ) external view override returns (uint256 usdAmountCharged, OrderFees.Data memory fees) {
         SpotMarketFactory.load().validateMarket(marketId);
 
@@ -160,7 +160,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function quoteSellExactIn(
         uint128 marketId,
-        uint synthAmount
+        uint256 synthAmount
     ) external view override returns (uint256 returnAmount, OrderFees.Data memory fees) {
         SpotMarketFactory.load().validateMarket(marketId);
 
@@ -178,7 +178,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function quoteSellExactOut(
         uint128 marketId,
-        uint usdAmount
+        uint256 usdAmount
     ) external view override returns (uint256 synthToBurn, OrderFees.Data memory fees) {
         SpotMarketFactory.load().validateMarket(marketId);
 
@@ -196,10 +196,10 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function sell(
         uint128 marketId,
-        uint synthAmount,
-        uint minUsdAmount,
+        uint256 synthAmount,
+        uint256 minUsdAmount,
         address referrer
-    ) external override returns (uint usdAmountReceived, OrderFees.Data memory fees) {
+    ) external override returns (uint256 usdAmountReceived, OrderFees.Data memory fees) {
         return sellExactIn(marketId, synthAmount, minUsdAmount, referrer);
     }
 
@@ -209,7 +209,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
     function sellExactIn(
         uint128 marketId,
         uint256 synthAmount,
-        uint minAmountReceived,
+        uint256 minAmountReceived,
         address referrer
     ) public override returns (uint256 returnAmount, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
@@ -232,7 +232,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
         // Burn after calculation because skew is calculating using total supply prior to fill
         SynthUtil.getToken(marketId).burn(msg.sender, synthAmount);
 
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
@@ -251,10 +251,10 @@ contract AtomicOrderModule is IAtomicOrderModule {
      */
     function sellExactOut(
         uint128 marketId,
-        uint usdAmount,
-        uint maxSynthAmount,
+        uint256 usdAmount,
+        uint256 maxSynthAmount,
         address referrer
-    ) external override returns (uint synthToBurn, OrderFees.Data memory fees) {
+    ) external override returns (uint256 synthToBurn, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
         spotMarketFactory.validateMarket(marketId);
 
@@ -272,7 +272,7 @@ contract AtomicOrderModule is IAtomicOrderModule {
         }
 
         SynthUtil.getToken(marketId).burn(msg.sender, synthToBurn);
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
