+++ markets/spot-market/contracts/modules/WrapperModule.sol
@@ -45,7 +45,7 @@ contract WrapperModule is IWrapperModule {
     function wrap(
         uint128 marketId,
         uint256 wrapAmount,
-        uint minAmountReceived
+        uint256 minAmountReceived
     ) external override returns (uint256 amountToMint, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
         Wrapper.Data storage wrapperStore = Wrapper.load(marketId);
@@ -79,7 +79,7 @@ contract WrapperModule is IWrapperModule {
             revert InsufficientAmountReceived(minAmountReceived, amountToMint);
         }
 
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
@@ -99,7 +99,7 @@ contract WrapperModule is IWrapperModule {
     function unwrap(
         uint128 marketId,
         uint256 unwrapAmount,
-        uint minAmountReceived
+        uint256 minAmountReceived
     ) external override returns (uint256 returnCollateralAmount, OrderFees.Data memory fees) {
         SpotMarketFactory.Data storage spotMarketFactory = SpotMarketFactory.load();
         Wrapper.Data storage wrapperStore = Wrapper.load(marketId);
@@ -129,7 +129,7 @@ contract WrapperModule is IWrapperModule {
         if (returnCollateralAmount < minAmountReceived) {
             revert InsufficientAmountReceived(minAmountReceived, returnCollateralAmount);
         }
-        uint collectedFees = config.collectFees(
+        uint256 collectedFees = config.collectFees(
             marketId,
             fees,
             msg.sender,
